third_party/blink/web_tests/fast/block/align-inverted-direction-expected.txt

PASS rtl_in_ltr.getBoundingClientRect().left is 0
PASS ltr_in_ltr.getBoundingClientRect().left is 0
PASS rtl_in_rtl.getBoundingClientRect().right is 700
PASS ltr_in_rtl.getBoundingClientRect().right is 700
PASS rtl_in_ltr_r.getBoundingClientRect().right is 700
PASS ltr_in_ltr_r.getBoundingClientRect().right is 700
PASS rtl_in_rtl_l.getBoundingClientRect().left is 0
PASS ltr_in_rtl_l.getBoundingClientRect().left is 0
PASS successfullyParsed is true

TEST COMPLETE
